requirements.txt
requests==2.31.0
beautifulsoup4==4.12.0
selenium==4.15.0
google-generativeai==0.3.0
sqlalchemy==2.0.0
python-dotenv==1.0.0
schedule==1.2.0
rich==13.7.0
